Candidate/Theme discovery
Uploading functions to read in candidates/themes provided tweet text.

Pass in text to return_candidates to receive a list comprised of a list
of candidates associated with that text (if any), a sentiment score, and
a subjectivity score.

Pass in text to return_themes to receive a list of themes associated
with that text (if any).